+++ webapp/views/utils.py
@@ -50,6 +50,8 @@ with DB.load_db(settings.BIODB_DB_PATH, settings.BIODB_CONF) as db:
                        (mandatory, value) in hsh_config.items() if not mandatory}
     optional2status["cog"] = optional2status["COG"]
     optional2status["ko"] = optional2status["KEGG"]
+    optional2status["module"] = optional2status["KEGG"]
+    optional2status["pathway"] = optional2status["KEGG"]
     optional2status["amr"] = optional2status["AMR"]
     optional2status["vf"] = optional2status["BLAST_vfdb"]
 
@@ -258,24 +260,32 @@ class DataTableConfig():
 
 class ResultTab():
 
-    def __init__(self, tabid, title, template, **kwargs):
+    def __init__(self, tabid, title, template, show_badge=False, badge=None, **kwargs):
         self.id = tabid
         self.title = title
         self.template = template
+        self.show_badge = show_badge
+        self.badge = badge
         for key, val in kwargs.items():
             setattr(self, key, val)
 
 
 class TabularResultTab(ResultTab):
 
-    def __init__(self, tabid, title, template, ordering=True, paging=True,
-                 export_buttons=True, colvis_button=False,
-                 display_index=False, **kwargs):
+    def __init__(self, tabid, title, template="chlamdb/result_table.html",
+                 ordering=True, paging=True, export_buttons=True,
+                 colvis_button=False, display_index=False,
+                 show_badge=False, **kwargs):
         self.data_table_config = DataTableConfig(
             table_id=tabid, ordering=ordering, paging=paging,
             export_buttons=export_buttons, colvis_button=colvis_button,
             display_index=display_index)
-        super(TabularResultTab, self).__init__(tabid, title, template, **kwargs)
+        if show_badge:
+            badge = len(kwargs["table_data"])
+        else:
+            badge = None
+        super(TabularResultTab, self).__init__(
+            tabid, title, template, show_badge=show_badge, badge=badge, **kwargs)
 
 
 def locusx_genomic_region(db, seqid, window):